+++ packages/components/src/internal/components/samples/FindSamplesByIdHeaderPanel.tsx
@@ -6,7 +6,8 @@ import { Button } from 'react-bootstrap';
 import { Alert } from '../base/Alert';
 import { FindByIdsModal } from '../navigation/FindByIdsModal';
 import { Section } from '../base/Section';
-import { getFindIdCountsByTypeMessage } from './actions';
+import { FIND_IDS_SESSION_STORAGE_KEY, SAMPLE_ID_FIND_FIELD, UNIQUE_ID_FIND_FIELD } from './constants';
+import { Utils } from '@labkey/api';
 
 interface HeaderPanelProps {
     loadingState: LoadingState,
@@ -16,8 +17,27 @@ interface HeaderPanelProps {
     onClearSamples: () => void,
 }
 
+function getFindIdCountsByTypeMessage() : string {
+    const findIds: string[] = JSON.parse(sessionStorage.getItem(FIND_IDS_SESSION_STORAGE_KEY))
+    if (!findIds) {
+        return undefined;
+    }
+
+    let numIdsMsg = '';
+    const numSampleIds = findIds.filter(id => id.startsWith(SAMPLE_ID_FIND_FIELD.storageKeyPrefix)).length;
+    const numUniqueIds = findIds.filter(id => id.startsWith(UNIQUE_ID_FIND_FIELD.storageKeyPrefix)).length;
+    if (numSampleIds) {
+        numIdsMsg += Utils.pluralize(numSampleIds, SAMPLE_ID_FIND_FIELD.nounSingular, SAMPLE_ID_FIND_FIELD.nounPlural);
+    }
+    if (numUniqueIds) {
+        numIdsMsg += (numIdsMsg ? ' and ' : '') + Utils.pluralize(numUniqueIds, UNIQUE_ID_FIND_FIELD.nounSingular, UNIQUE_ID_FIND_FIELD.nounPlural);
+    }
+    return numIdsMsg;
+}
+
 export const FindSamplesByIdHeaderPanel: FC<HeaderPanelProps> = memo((props) => {
     const [showFindModal, setShowFindModal] = useState<boolean>(false);
+
     const { loadingState, listModel, onFindSamples, onClearSamples, missingIds } = props;
 
     const numIdsMsg = getFindIdCountsByTypeMessage();
@@ -25,7 +45,6 @@ export const FindSamplesByIdHeaderPanel: FC<HeaderPanelProps> = memo((props) => 
     const onAddMoreSamples = useCallback(() => {
         setShowFindModal(true);
     }, []);
-
 
     const onCancelAdd = useCallback(() => {
         setShowFindModal(false)
@@ -47,7 +66,7 @@ export const FindSamplesByIdHeaderPanel: FC<HeaderPanelProps> = memo((props) => 
         foundSamplesMsg = (
             <div className="bottom-spacing">
                 <i className="fa fa-check-circle find-samples-success"/>{' '}
-                <span>Found {listModel.rowCount} samples matching {numIdsMsg}.</span>
+                <span>Found {Utils.pluralize(listModel.rowCount, 'sample', 'samples')} matching {numIdsMsg}.</span>
             </div>
         );
     }
@@ -106,7 +125,7 @@ export const SamplesNotFoundMsg: FC<{missingIds: {[key: string]: string[]}}> = m
             <div className="bottom-spacing">
                 <span className="find-samples-warning"><i className="fa fa-exclamation-circle"/> </span>
                 <span>
-                    Couldn't locate {allIds.length} samples{' '}
+                    Couldn't locate {Utils.pluralize(allIds.length, 'sample', 'samples')}{' '}
                     <a className="find-samples-warning-toggle" onClick={toggleShowIdAlert}>
                         {showIds ?
                             <>Hide <i className="fa fa-caret-down" aria-hidden="true"/></>:
