+++ packages/components/src/theme/navbar.scss
@@ -379,10 +379,6 @@
 
   .navbar__input-group {
       width: 320px;
-
-      .navbar__search-input {
-          border-radius: 4px;
-      }
   }
   .navbar__find-and-search-button {
       border: none;
